Update fix issues/5792

+++ src/Eccube/Controller/Admin/Setting/Shop/DeliveryController.php
@@ -282,7 +282,7 @@ class DeliveryController extends AbstractController
                 return $this->redirectToRoute('admin_setting_shop_delivery_edit', ['id' => $Delivery->getId()]);
             } else {
                 foreach ($Delivery->getDeliveryTimes() as $deliveryTime) {
-                    if ($deliveryTime->getDeliveryTime() == null) {
+                    if ($deliveryTime->getDeliveryTime() == null && $deliveryTime->getId() != null) {
                         $this->entityManager->detach($deliveryTime);
                         $deliveryTime->setDeliveryTime($this->deliveryTimeRepository->find($deliveryTime->getId())->getDeliveryTime());
                     }
